Resolve issues with "/SYM64/" support changes
- missing initialization of w_sz
- type and size issues
- support for reading "/SYM64/" archives
still requires support from libarchive

+++ usr.bin/ar/ar.h
@@ -100,7 +100,7 @@ struct bsdar {
 	/*
 	 * Fields for the archive symbol table.
 	 */
-	uint32_t	  s_cnt;	/* current number of symbols. */
+	uint64_t	  s_cnt;	/* current number of symbols. */
 	uint64_t	 *s_so;		/* symbol offset table. */
 	uint64_t	  s_so_max;     /* Maximum symbol offset (for determining which ar format to use)*/
 	size_t		  s_so_cap;	/* capacity of so table buffer. */
